Move the four HOME screens out of EcuLab.jsx
HOME becomes four files under src/ui/screens/dash — live, stats, health and
learn — each reading the store for itself and each owning its own accordion
section. The markup is unchanged: this moves it, it does not re-section
anything (that is #83).

LiveScreen is the reason for the split. `session.live` is rewritten twenty
times a second by LIVE_STEP, and every one of those writes has to re-render
the live panel. Keeping `live` read inside that one file means the re-render
stops there instead of passing through a HOME-level parent that would drag
the other three sections with it. LearnScreen — 16 static articles, the
largest block of markup on the tab — reads no store at all and is memoised,
so it is skipped entirely at 20 Hz rather than walked twenty times a second
to produce the same output. That needs a stable `onToggle`, which is why the
HOME toggle now reads the current section from a ref.

Three components are shared by tabs this task does not touch, so they move to
primitives/ rather than into a screen: BuildSection, ExpandableInfo and
DialMark. They keep their inline styles verbatim — BuildSection's especially,
since routing-shell.test.jsx reads its inline maxHeight as the only
DOM-visible difference between an open section and a closed one.

Derivations were decided one at a time. `tachFullScaleRpm` (dyno tach too),
`scores` (DYNO scorecard too), `overallHealth` (header bar too) and
`needsMafRecal` (TUNE > ECU too) each have more than one consumer, so they
stay in the shell and are passed in; recomputing them in a screen would be
two answers to one question. LiveGauge and TrimBar had exactly one consumer
and moved into LiveScreen with it.

button-call-sites.test.jsx now scans the shell AND every screen file rather
than EcuLab.jsx alone, with the floor raised from 18 to 23 — the total across
the glob. Extracting a tab moves call sites, it does not delete them, so the
number this guards must go up, not down.
